NPB refactoring
