+++ ydb/core/blobstorage/ddisk/persistent_buffer_space_allocator.h
@@ -1,18 +1,11 @@
 #pragma once
 
 #include "defs.h"
+#include "persistent_buffer_header.h"
 
 #include <queue>
 
 namespace NKikimr::NDDisk {
-
-    struct TPersistentBufferSectorInfo {
-        ui64 ChunkIdx : 32;
-        ui64 SectorIdx : 16;
-        ui64 HasSignatureCorrection : 1;
-        ui64 Reserved : 15;
-        ui64 Checksum : 64;
-    };
 
     class TPersistentBufferSpaceAllocator {
     protected:
